Set up flows in MainViewModel.kt

+++ app/src/main/java/com/terranullius/clean_cache_mvvm/business/data/cache/abstraction/UserCacheDataSource.kt
@@ -7,7 +7,7 @@ interface UserCacheDataSource {
 
     suspend fun insertUser(user: User): Long
 
-    suspend fun deleteUser(user: User): Long
+    suspend fun deleteUser(user: User): Int
 
     suspend fun getSavedUsers(): DataState
 }
